Change p/invoke method signature.

+++ StarryEyes.Nightmare/Windows/WindowNativeExtension.cs
@@ -11,7 +11,7 @@ namespace StarryEyes.Nightmare.Windows
             var helper = new WindowInteropHelper(window);
             var wpl = new WINDOWPLACEMENT();
             wpl.length = Marshal.SizeOf(wpl);
-            WinApi.GetWindowPlacement((int)helper.Handle, ref wpl);
+            WinApi.GetWindowPlacement(helper.Handle, ref wpl);
             return new Rect(wpl.rcNormalPosition.left, wpl.rcNormalPosition.top,
                 wpl.rcNormalPosition.right - wpl.rcNormalPosition.left,
                 wpl.rcNormalPosition.bottom - wpl.rcNormalPosition.top);
@@ -26,7 +26,7 @@ namespace StarryEyes.Nightmare.Windows
             wpl.rcNormalPosition.right = (int)placement.Right;
             wpl.rcNormalPosition.top = (int)placement.Top;
             wpl.rcNormalPosition.bottom = (int)placement.Bottom;
-            WinApi.SetWindowPlacement((int)helper.Handle, ref wpl);
+            WinApi.SetWindowPlacement(helper.Handle, ref wpl);
         }
     }
 }
